'Migrated' outside qt
We now do the includes relative to each file. We really tried not to but I dont
undestand how the dozens of tools that are need to fucking build a simple c++
project really work.
We tried to add the source directories to cmake but it didnt work, it didnt throw
an error or anything either so fucking idea of what is going on.
BUT right now it compiles.